Split expression method call checking
