+++ src/typechecker/expressions.rs
@@ -5,6 +5,7 @@ mod aggregate_support;
 mod call_support;
 mod call_validation;
 mod control_flow_support;
+mod method_call_support;
 
 use crate::ast::expressions::StringPart;
 use crate::ast::typed::*;
